+++ lib/aasm/instance_base.rb
@@ -30,8 +30,15 @@ module AASM
       AASM::Localizer.new.human_state_name(@instance.class, current_state)
     end
 
-    def states
-      @instance.class.aasm.states
+    def states(options={})
+      if options[:permissible]
+        # ugliness level 1000
+        transitions = @instance.class.aasm.events.values.map {|e| e.transitions_from_state(current_state) }
+        tos = transitions.map {|t| t[0] ? t[0].to : nil}.flatten.compact.map(&:to_sym).uniq
+        @instance.class.aasm.states.select {|s| tos.include?(s.name.to_sym)}
+      else
+        @instance.class.aasm.states
+      end
     end
 
     # QUESTION: shouldn't events and permissible_events be the same thing?